Using entries in the form of order-title
Using entries in the form of order-title instead of order-name.  As
happen in cases such as hosts defined as 'eth0:0.0.0.0/0', the '/' will
make the first portion to be understood as a folder.  Making use of the
resource title as part of the file name and leaving the resource name to
the content can prevent this error.