Remove 'require augeas' (#4396)
Because the puppet 'require' keyword actually instantiates a class,
having 'require augeas' in base.pp means that you cannot instantiate the
augeas class anywhere else in your manifests, for example with some
optional parameters to the class. If you do, you will get a duplicate
definition error.

The README already says that the augeas module is required. It seems
better that this is managed outside of the module, allowing for
class parameter flexibility.